+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 ~~~~~~~~~~~~~~~~~~~
 
 - Unregistered extensions are now supported by the profile builder. The ``extension_versions`` argument to ``ProfileBuilder`` can be a list of extensions' metadata URLs, base URLs and/or download URLs.
+- Add an ``annotate`` parameter to ``ProfileBuilder`` to annotate all definitions and properties with extension names.
 
 0.0.11 (2019-06-26)
 ~~~~~~~~~~~~~~~~~~~